requirements.txt
certifi==2017.11.5
chardet==3.0.4
click==6.7
coreapi==2.3.3
coreschema==0.0.4
Django==1.11
django-debug-toolbar==1.9.1
django-rest-swagger==2.1.2
djangorestframework==3.7.7
djangorestframework-jsonapi==2.3.1
drf-openapi==1.3.0
idna==2.6
inflection==0.3.1
itypes==1.1.0
Jinja2==2.10
MarkupSafe==1.0
openapi-codec==1.3.2
pytz==2017.3
requests==2.18.4
simplejson==3.13.2
six==1.11.0
sqlparse==0.2.4
uritemplate==3.0.0
urllib3==1.22
